Move the pipelined replay test to test/regression/issue/40350.test.ts
The bug is a true regression (1.3.14 worked, 1.4.0 broke), and
node-http.test.ts has environment-dependent failures that are unrelated
to this change.
